Moving implemented.
Dinosaurs follow after the survivor very well. Going to need to tweak
the settings a little, but it's at a playable level.

Also reduced the collision boxes of all dinosaurs, so they can go
through gaps in trees. When I make bigger ones, they'll not be able to
do the same.

Maybe I should put in a death-timer on dinosaurs? So they can't live for
longer than 10 minutes or something. I need a way to stop them spawning
in gaps they can't escape from.