Finished Pistol
It doesn't sound like much, but it lays the foundations for all of the
firearms out there.

I also brought in an external library for displaying text for hit
chance, and any misses. I modified it for my needs.

I've got to add the skills for training/bonuses, and longer cooldown for
jammed guns.

Not to mention, I have to actually test this.