Progress on specialization structs
Made upgrades for specializations, which will be upgraded to choose a
class.

I'm pretty intimidated by trying to implement this well, but have
decided on each survivor unit having a survivor struct. The survivor
struct contains general details about the survivor, as well as 4 more
structs - one for each tech tree.

I wanted to make the specialization struct extend the survivor one, but
there might be issues with casting, and will restrict units to being one
class (if I ever change my mind in the future).

There are proper algorithms for dino following now.

Code is certainly broken, and is not actually in map atm.