Mostly completed code for evacuation (end game)
At the end of the game, there is now an evacuation zone which is pinged, and the
survivors must all make it there. There is a chopper waiting, and the game announces
who made it there.

Completely removed old spawn and move triggers.

Started taking a look at the Garage. The code for swapping items sets to sell will have
to be done using Jass.

I also did a little more terraining.