Added swap function for shops
Wrote some absolutely brilliant code here.

It sets the items in the shop based on the amount of mana that the shop has.
It toggles between 1 and 2, but it could have any number!

This could be really useful for other swappable things.

At the moment the items for the Mechanic are very in-limbo,
but I'll work on them next - correct naming, data values, etc,
and maybe logical increments. Should change everything to
object oriented.